Significantly improve speed of sourcing files
This also includes processing configuration on startup and :restart.

ui_sb_msg() is quite heavyweight in that it tasks curses with redrawing
essentially the whole screen every time.  Invoking it after processing
every line is what caused awful performance (the delay was noticeable
even on processing vifmrc of 500 comments with some empty lines!).