Don't handle SIGCHLD
It makes it harder to use system(), popen() or start processes using
fork as well as waiting for them using waitpid().  From now on children
are ripped synchronously by polling.

SIGCHLD handler was always there, but it mostly made the life harder.
In particular it doesn't play well with bg_job_t as signal handler might
want to interrupt thread after acquiring status lock of the job and then
try to acquire it again from the handler to produce a deadlock.