Fix implicit selection being lost (again)
356d95eb515cde7b12c (Fix issues with implicit file selection) was fixing
a problem demonstrated by b823d223396773993a6 (Preserve pending_marking
flags by ma_expand()) and effectively reverted that commit.

This change has a somewhat different take on the fix.  The underlying
issue is that expanding is linked to consuming marking, so make it
possible to do expansion without touching marking (using selection
only).

This keeps saving/restoring of `preserve_marking` fields in macros.c and
resetting them in `event_loop()` just in case, because it's hard to make
sure that marking is never carried over to the next operation in general.

Related to #608 and #604 on GitHub.

Fixes #763 on GitHub.