Fix assertion failure on processing some mappings
Those with RHS of the form "{mapping with selector}{something else}".

The issue was that "mapped" flag was dropped for the
continuation ({something else} part) causing it to affect counter of
processed input keys while event loop asserts for its value not be
larger than the input buffer.

Thanks to dalvand.

See:

https://q2a.vifm.info/1125/remap-da-to-da
https://q2a.vifm.info/1144/workaround-to-automatically-retrieve-selection-after-open?show=1158#a1158