Fix selection issues after on switching av to v
Selection created prior to running av in normal mode was showing up
after switching to a regular visual mode via v key and moving cursor
over it.

The original selection is remembered so it can be restored in av mode,
but it wasn't cleared on transition to v mode, which caused the issue.