Fix a "benign" data race on ga/gA
It's benign, because background thread was only reading global data
which is always there.  Nothing bad as a result of that.

Something to think about in relation to this: a queue of functions to
run in main thread that can be populated by background threads;
alternatively, on-exit code for background threads which is run by main
thread.  Either would allow doing UI updates when background thread is
over.